# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM PAY API",
        "description": "API de paiement mobile OM PAY avec authentification OTP et gestion de comptes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://ompay-isuf.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "description": "Crée un compte utilisateur, génère un QR code et envoie un OTP par email",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "telephone",
                                    "email"
                                ],
                                "properties": {
                                    "nom": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "example": "Diop"
                                    },
                                    "prenom": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string",
                                        "example": "Amadou"
                                    },
                                    "telephone": {
                                        "description": "Numéro de téléphone (9 chiffres)",
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "email": {
                                        "description": "Adresse email de l'utilisateur",
                                        "type": "string",
                                        "example": "amadou.diop@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur créé avec succès. Vérifiez votre email pour le code OTP."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "uuid-string"
                                                        },
                                                        "nom": {
                                                            "type": "string",
                                                            "example": "Diop"
                                                        },
                                                        "prenom": {
                                                            "type": "string",
                                                            "example": "Amadou"
                                                        },
                                                        "telephone": {
                                                            "type": "string",
                                                            "example": "771234567"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "amadou.diop@example.com"
                                                        },
                                                        "is_verified": {
                                                            "type": "boolean",
                                                            "example": false
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de l'inscription"
                    }
                }
            }
        },
        "/auth/verify-otp": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérification du code OTP",
                "description": "Valide le code OTP reçu par email et définit un PIN temporaire (0000)",
                "operationId": "verifyOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "otp"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone",
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "otp": {
                                        "description": "Code OTP à 6 chiffres",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vérification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vérification réussie. Votre PIN temporaire est 0000. Veuillez le changer immédiatement."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "uuid-string"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "client"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "bearer-token-string"
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "refresh-token-string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "temporary_pin": {
                                                    "type": "string",
                                                    "example": "0000"
                                                },
                                                "requires_pin_change": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Code OTP invalide ou expiré"
                    }
                }
            }
        },
        "/auth/set-pin": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Définition du PIN définitif",
                "description": "Permet à l'utilisateur de définir son PIN définitif après la première connexion avec le PIN temporaire",
                "operationId": "setPin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_pin"
                                ],
                                "properties": {
                                    "code_pin": {
                                        "description": "Nouveau code PIN définitif à 4 chiffres",
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PIN définitif défini avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PIN définitif défini avec succès. Vous pouvez maintenant utiliser votre compte normalement."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "PIN déjà défini ou invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion utilisateur",
                "description": "Invalide le token d'accès actuel",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}